+++ History.md
@@ -1,3 +1,19 @@
+0.7.2 / 2014-05-12
+==================
+  * Dependency update to allow pantry v0.7.x
+  * Dependency update to allow request v2.36.x
+  * Dependency update to allow express v3.10.x
+  * Dependency update to allow hjs v0.0.6
+  * Dependency update to allow hogan.js v3.0.x
+  * Dependency update to allow run v1.2.x
+  * Dependency update to allow mocha v1.20.x
+  * Dependency update to allow should v4.0.x
+  * Examples - added parameter and example help for dutchie timbit
+  * Added dependency for should-http v0.0.2 augments should v4.x
+  * Updated node engine support from 0.8.x to up to 0.10.x
+  * Updated npm to 1.x 
+  * Changed test to require '-c' or '--coffee' to use coffee-script
+
 0.7.1 / 2013-07-16
 ==================
   * Dependency update to allow pantry v0.5.x